+++ cmd/substreams/protogen.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"path/filepath"
 
-	"github.com/streamingfast/substreams/tools"
-
 	"github.com/spf13/cobra"
 	"github.com/streamingfast/cli"
 	"github.com/streamingfast/substreams/codegen"
@@ -44,14 +42,11 @@ func runProtogen(cmd *cobra.Command, args []string) error {
 	excludePaths := mustGetStringArray(cmd, "exclude-paths")
 	generateMod := mustGetBool(cmd, "generate-mod-rs")
 
-	manifestPathRaw := ""
+	manifestPath := ""
 	if len(args) == 1 {
-		manifestPathRaw = args[0]
+		manifestPath = args[0]
 	}
-	manifestPath, err := tools.ResolveManifestFile(manifestPathRaw)
-	if err != nil {
-		return fmt.Errorf("resolving manifest: %w", err)
-	}
+
 	manifestReader := manifest.NewReader(manifestPath, manifest.SkipSourceCodeReader(), manifest.SkipModuleOutputTypeValidationReader())
 
 	if manifestReader.IsLocalManifest() && !filepath.IsAbs(outputPath) {