+++ cmd/substreams/run.go
@@ -7,18 +7,15 @@ import (
 	"strconv"
 	"strings"
 
-	"go.uber.org/zap"
-
-	"github.com/streamingfast/substreams/tools/test"
-
 	"github.com/spf13/cobra"
 	"github.com/streamingfast/cli"
-
 	"github.com/streamingfast/substreams/client"
 	"github.com/streamingfast/substreams/manifest"
 	pbsubstreamsrpc "github.com/streamingfast/substreams/pb/sf/substreams/rpc/v2"
 	"github.com/streamingfast/substreams/tools"
+	"github.com/streamingfast/substreams/tools/test"
 	"github.com/streamingfast/substreams/tui"
+	"go.uber.org/zap"
 )
 
 func init() {
@@ -56,24 +53,16 @@ var runCmd = &cobra.Command{
 
 func runRun(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
-	outputMode := mustGetString(cmd, "output")
 
-	manifestPath := ""
-	var err error
-	if len(args) == 2 {
-		manifestPath = args[0]
-		args = args[1:]
+	var manifestPath, outputModule string
+	if len(args) == 1 {
+		outputModule = args[0]
 	} else {
-		if cli.DirectoryExists(args[0]) || cli.FileExists(args[0]) || strings.Contains(args[0], ".") {
-			return fmt.Errorf("parameter entered likely a manifest file, don't forget to include a '<module_name>' in your command")
-		}
-
-		// At this point, we assume the user invoked `substreams run <module_name>` so we `ResolveManifestFile` using the empty string since no argument has been passed.
-		manifestPath, err = tools.ResolveManifestFile("")
-		if err != nil {
-			return fmt.Errorf("resolving manifest: %w", err)
-		}
+		manifestPath = args[0]
+		outputModule = args[1]
 	}
+
+	outputMode := mustGetString(cmd, "output")
 
 	manifestReader := manifest.NewReader(manifestPath)
 	pkg, err := manifestReader.Read()
@@ -112,8 +101,6 @@ func runRun(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return fmt.Errorf("creating module graph: %w", err)
 	}
-
-	outputModule := args[0]
 
 	startBlock, readFromModule, err := readStartBlockFlag(cmd, "start-block")
 	if err != nil {