+++ manifest/reader_test.go
@@ -15,6 +15,9 @@ import (
 )
 
 func TestReader_Read(t *testing.T) {
+	absolutePathToInferredManifest, err := filepath.Abs("testdata/inferred_manifest")
+	require.NoError(t, err)
+
 	absolutePathToDep2, err := filepath.Abs("testdata/dep2.yaml")
 	require.NoError(t, err)
 
@@ -30,8 +33,11 @@ func TestReader_Read(t *testing.T) {
 	defer remoteServer.Close()
 
 	type args struct {
-		env            map[string]string
-		validateBinary bool
+		// If nil, the input is taken from the name
+		input            *string
+		env              map[string]string
+		validateBinary   bool
+		workingDirectory string
 	}
 
 	tests := []struct {
@@ -43,6 +49,41 @@ func TestReader_Read(t *testing.T) {
 		{
 			"bare_minimum.yaml",
 			args{},
+			&pbsubstreams.Package{
+				Version:    1,
+				ProtoFiles: readSystemProtoDescriptors(t),
+				Modules:    &pbsubstreams.Modules{},
+				PackageMeta: []*pbsubstreams.PackageMetadata{
+					{
+						Name:    "test",
+						Version: "v0.0.0",
+					},
+				},
+			},
+			require.NoError,
+		},
+		{
+			"from_folder",
+			args{},
+			&pbsubstreams.Package{
+				Version:    1,
+				ProtoFiles: readSystemProtoDescriptors(t),
+				Modules:    &pbsubstreams.Modules{},
+				PackageMeta: []*pbsubstreams.PackageMetadata{
+					{
+						Name:    "test",
+						Version: "v0.0.0",
+					},
+				},
+			},
+			require.NoError,
+		},
+		{
+			"empty_input",
+			args{
+				input:            new(string),
+				workingDirectory: absolutePathToInferredManifest,
+			},
 			&pbsubstreams.Package{
 				Version:    1,
 				ProtoFiles: readSystemProtoDescriptors(t),
@@ -228,9 +269,6 @@ func TestReader_Read(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			manifestPath, err := filepath.Abs(filepath.Join("testdata", tt.name))
-			require.NoError(t, err)
-
 			for envKey, envValue := range tt.args.env {
 				t.Setenv(envKey, envValue)
 			}
@@ -240,8 +278,23 @@ func TestReader_Read(t *testing.T) {
 				readerOptions = append(readerOptions, SkipSourceCodeReader())
 			}
 
+			var manifestPath string
+			if tt.args.input != nil {
+				manifestPath = *tt.args.input
+			} else {
+				var err error
+				manifestPath, err = filepath.Abs(filepath.Join("testdata", tt.name))
+				require.NoError(t, err)
+			}
+
 			r := NewReader(manifestPath, readerOptions...)
-			got, err := r.Read()
+
+			workingDir := ""
+			if tt.args.workingDirectory != "" {
+				workingDir = tt.args.workingDirectory
+			}
+
+			got, err := r.read(workingDir)
 			tt.assertion(t, err)
 			assertProtoEqual(t, tt.want, got)
 		})